+++ docs/guide/changelog.md
@@ -1,5 +1,15 @@
 # Changelog
-## 2.7.36 (2025-06-16)
+
+
+## 2.7.39 (2025-06-19)
+
+- Feat
+  - DatePicker: Added `format` prop to allow customization of date format in the date picker (e.g., 'MM-DD-YYYY', 'YYYY-MM-DD', 'MM/DD/YYYY')
+  - DatePicker: Fixed text color for today's date when selected to be white instead of brand color
+  
+  ([#588f2034](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/588f2034f7df6212a93c89a6891d0101bd53f0cf?refName=refs/heads/2025/FEATURE/BEN/DATEPICKER) by @bsolis)
+
+## 2.7.38 (2025-06-17)
 
 - Enhancement
   - Dropdown component improvements for object values:
@@ -19,6 +29,7 @@
 
 - Feat
   - Commit ddf1ee3b: added calendar empty state ([#ddf1ee3b](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/ddf1ee3b44d05dd382c1272961e2d51744bac4f1?refName=refs/heads/2025/feat/andrea/empty-state-calendar) by @adestajo)
+
 
 ## 2.7.35 (2025-06-04)
 
@@ -127,7 +138,7 @@
 ## 2.7.15 (2025-05-13)
 
 - Fix
-  - Fix switch issue clickable label ([#30182d0f](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/30182d0f6039794e4532d452649463b048859b99?refName=refs/heads/2025/FIX/BEN/SWITCH_CHECK_LABEL) by @bsolis)
+  - Fix switch issue clickable label  ([#30182d0f](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/30182d0f6039794e4532d452649463b048859b99?refName=refs/heads/2025/FIX/BEN/SWITCH_CHECK_LABEL) by @bsolis)
 
 ## 2.7.14 (2025-05-09)